fixed warnings

+++ coreservices/partsrelationshipservice/prs-api/src/main/resources/db/migration/V1_12__part_relationship.sql
@@ -1,2 +1,2 @@
 ALTER TABLE public.part_relationship
-    ADD COLUMN version integer NOT NULL DEFAULT 0;
+    ADD COLUMN version bigint NOT NULL DEFAULT 0;
